update
At this point I'm having some pointer difficulties.  I was referencing
one copy og _box as a static member of the Game class, but getting
different addresses for beased on which class it was in.  so I made it a
global, and defined a getBox() method in the Game.h to call it.  Now I'm
seeinf other values being overwritten with memory locations valutes.
Like flagId.  So this is a work in progress.  that said it does install
to the NodeMCU V3 (LoLin)D3-D8 are the 5 input buttons.  I'm not at the
point where I can actually readuser input,  It is coded, I just need to
get the pointers right.