Version 0.0.1 Complete, working Coin Flip
Yes!  The initial framework is done and working.  CoinFlip works.  Of
course all the stages for all the flags are currently Coin Toss, but who
cares.  It works.  Now I can add in more stages and more flags.  I need
to add in the RTC and some more framework.  The LCD is not in use yet,
you need the Serial console.