Update
I have all of Flag one complete.  All three stages are done, with the
exception of a ghost bug.  On the third stage if I press a button more
than once it can be mistaken for another button.  I have implemented
code that sets a bool when a button is pressed and it doen;t read it.
so the digitalRead seems to pick up on the LOW from a neighboring pin.
Needs Research.